% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BrandCode.r
\docType{data}
\name{BrandCode}
\alias{BrandCode}
\title{Brands listed on the Tokyo Stock Exchange}
\format{
A data frame with 4170 rows and 4 variables:
\describe{
\item{code}{brand code}
\item{name}{company name}
\item{category}{market and product category}
\item{indust17}{17 types of industries}
}
}
\source{
\url{
https://www.jpx.co.jp/markets/statistics-equities/misc/01.html}
}
\usage{
BrandCode
}
\description{
A dataset containing brands
listed on the Tokyo Stock Exchange
as of the end of December 2021 published by the JPX.
}
\keyword{datasets}
