% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_detect.r
\name{code_detect}
\alias{code_detect}
\title{Search for Stock Code by Company Name}
\usage{
code_detect(firm)
}
\arguments{
\item{firm}{a character vector. Specify company names}
}
\value{
return a list object.
The "code" element contains a vector of stock codes.
The "info" element contains the matched data.
}
\description{
It is a function to search the brand code
from the company name.
The stock code data uses the list of brands
listed on the Tokyo Stock Exchange
as of the end of December 2021 published by the JPX (
\url{https://www.jpx.co.jp/markets/statistics-equities/misc/01.html}).
The data used inside the function
can be confirmed by \verb{data(Brand Code)}.
Concatenate vectors containing company names with \code{|}
and return the matched company
with the \code{str_detect} function of the {stringr} package.
}
