% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_onefirm.r
\name{page_count}
\alias{page_count}
\title{Calculate The Number of Pages of Stock Price Data of A Certain Brand}
\usage{
page_count(code, start_date, end_date, datatype)
}
\arguments{
\item{code}{numeric. Specify a brand code}

\item{start_date}{numeric.
Specify the start date of the stock price data you want to acquire
in the format of yyyymmdd.}

\item{end_date}{numeric.
Specify the end date of the stock price data you want to acquire
in the format of yyyymmdd.}

\item{datatype}{character. Specify the type of stock price data.
If \verb{datatype = "d", get daily data. If }datatype = "w", get weekly data.
If `datatype = "m", get monthly data.}
}
\description{
Calculate The Number of Pages of Stock Price Data of A Certain Brand
}
