% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_onefirm.r
\name{scrape_onefirm}
\alias{scrape_onefirm}
\title{Web Scraping of Stock Price Data for One Brand}
\usage{
scrape_onefirm(code, start_date, end_date, datatype)
}
\arguments{
\item{code}{numeric. Specify a brand code}

\item{start_date}{numeric.
Specify the start date of the stock price data you want to acquire
in the format of yyyymmdd.}

\item{end_date}{numeric.
Specify the end date of the stock price data you want to acquire
in the format of yyyymmdd.}

\item{datatype}{character. Specify the type of stock price data.
If \verb{datatype = "d", get daily data. If }datatype = "w", get weekly data.
If `datatype = "m", get monthly data.}
}
\value{
A tibble data containing stock price data
}
\description{
Calculate the number of pages required
to get all stock price data for one brand,
and execute \code{scrape_onepage} for each page using the loop syntax.
At this time, in order to avoid the load on the server,
after finishing scraping for one page,
wait one second and execute scraping for the next page.
Use the \code{bind_rows} of the {dplyr} package
o connect the stock price data on each page in the row direction.
}
\examples{
# Acquire 2014 daily stock price data of Sony Group (brand code 6758)
scrape_onefirm(6785, 20140101, 20141231, "d")

}
