% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_onefirm.r
\name{scrape_onepage}
\alias{scrape_onepage}
\title{Web Scraping of Stock Price Data}
\usage{
scrape_onepage(code, start_date, end_date, datatype, page)
}
\arguments{
\item{code}{numeric. Specify a brand code}

\item{start_date}{numeric. Specify the start date of the stock price data you want to acquire in the format of yyyymmdd.}

\item{end_date}{numeric. Specify the end date of the stock price data you want to acquire in the format of yyyymmdd.}

\item{datatype}{character. Specify the type of stock price data. If \verb{datatype = "d", get daily data. If }datatype = "w", get weekly data. If `datatype = "m", get monthly data.}

\item{page}{numeric. Specify the page number of the stock price data you want to acquire.}
}
\value{
A tibble data containing stock price data
}
\description{
It is a function to scrape stock price data
on a specific page of a certain brand from Yahoo! Finance.
This function is the backbone of this package,
and other functions are wrapper functions for this function.
}
